requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
alembic==1.17.1
    # via gco-pkm-llm (pyproject.toml)
annotated-types==0.7.0
    # via pydantic
anthropic==0.72.0
    # via gco-pkm-llm (pyproject.toml)
anyio==4.11.0
    # via
    #   anthropic
    #   httpx
bcrypt==5.0.0
    # via gco-pkm-llm (pyproject.toml)
blinker==1.9.0
    # via flask
certifi==2025.10.5
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
click==8.3.0
    # via flask
deprecated==1.3.0
    # via limits
distro==1.9.0
    # via anthropic
docstring-parser==0.17.0
    # via anthropic
flask==3.1.2
    # via
    #   gco-pkm-llm (pyproject.toml)
    #   flask-limiter
flask-limiter==4.0.0
    # via gco-pkm-llm (pyproject.toml)
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via anthropic
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
jiter==0.11.1
    # via anthropic
limits==5.6.0
    # via flask-limiter
mako==1.3.10
    # via alembic
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via
    #   flask
    #   jinja2
    #   mako
    #   werkzeug
mdurl==0.1.2
    # via markdown-it-py
ordered-set==4.1.0
    # via flask-limiter
packaging==25.0
    # via limits
psycopg2-binary==2.9.11
    # via gco-pkm-llm (pyproject.toml)
pydantic==2.12.3
    # via anthropic
pydantic-core==2.41.4
    # via pydantic
pygments==2.19.2
    # via rich
pyjwt==2.10.1
    # via gco-pkm-llm (pyproject.toml)
python-dotenv==1.2.1
    # via gco-pkm-llm (pyproject.toml)
pyyaml==6.0.3
    # via gco-pkm-llm (pyproject.toml)
requests==2.32.5
    # via gco-pkm-llm (pyproject.toml)
rich==14.2.0
    # via flask-limiter
sniffio==1.3.1
    # via
    #   anthropic
    #   anyio
sqlalchemy==2.0.44
    # via
    #   gco-pkm-llm (pyproject.toml)
    #   alembic
typing-extensions==4.15.0
    # via
    #   alembic
    #   anthropic
    #   anyio
    #   flask-limiter
    #   limits
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
urllib3==2.5.0
    # via requests
werkzeug==3.1.3
    # via flask
wrapt==2.0.0
    # via deprecated
